// src/app/shared/model/vendor/res-get-list-vendor-attr.model.ts
export class resGetListVendorAttr {
    ListVendorAttrContent: any[]; // Atau sesuai dengan tipe yang sesuai
    Jurisdiction: any; // Atau sesuai dengan tipe yang sesuai
    HeaderObj: {
      ResponseTime: string;
      StatusCode: string;
      Message: string;
      ErrorMessages: any; // Atau sesuai dengan tipe yang sesuai
    };
    StatusCode: string;
    Message: string;
    ErrorMessages: any; // Atau sesuai dengan tipe yang sesuai
    RowVersion: any; // Atau sesuai dengan tipe yang sesuai
  }